@@ -1,11 +1,11 @@
 module javaDebugDaemon
 
-go 1.17
+go 1.19
 
 require (
 	github.com/gin-gonic/gin v1.7.7
 	github.com/golang/protobuf v1.4.0 // indirect
-	golang.org/x/sys v0.0.0-20200420163511-1957bb5e6d1f // indirect
+	golang.org/x/sys v0.2.0 // indirect
 	k8s.io/api v0.17.0
 	k8s.io/client-go v0.17.0
 )
